requirements.txt
fastapi
uvicorn
requests
numpy==1.23.0
faiss-cpu
PyMuPDF
nltk
sentence-transformers
langdetect
google-generativeai
beautifulsoup4
httpx==0.24.1
scikit-learn==1.3.0
rank_bm25==0.2.2